requirements.txt
jinja2
groq
python-dotenv
sentence-transformers
scikit-learn
